Попытка2

+++ src/main/java/com/company/Calculator.java
@@ -7,7 +7,6 @@ public class Calculator {
     public int add(int a, int b) {
 
         return a + b;
-
     }
 
     public int minus(int a, int b) {
@@ -21,7 +20,7 @@ public class Calculator {
         return c;
     }
 
-    public double division(int a, int b) {
+    public int division(int a, int b) {
 
         return a / b;
     }
