+++ src/main/java/com/company/Main.java
@@ -19,7 +19,7 @@ public class Main {
         int result = calculator.add(a, b);
         int result1 = calculator.minus(a,b);
         long result2 = calculator.multiply(a,b);
-        double result3 = calculator.division(a,b);
+        int result3 = calculator.division(a,b);
 
         System.out.println(result);
         System.out.println(result1);
